+++ src/MetaheuristicsPlatform/Algorithms/Multiobjective/PesaII/PesaIIReferences.cs
@@ -1,0 +1,10 @@
+using MetaheuristicsPlatform.Classification;
+namespace MetaheuristicsPlatform.Algorithms.Multiobjective.PesaII;
+public static class PesaIIReferences
+{
+    public static ScientificReference CorneJerramKnowlesOates2001 { get; } =
+        new("Corne, Jerram, Knowles & Oates", 2001,
+            "PESA-II: Region-Based Selection in Evolutionary Multiobjective Optimization",
+            "GECCO 2001, 283-290",
+            "10.5555/2955239.2955289");
+}